Add webtiles popup UI stack
This commit adds an additional webtiles client API to show popups.
Unlike the existing show_dialog() and hide_dialog() functions, the new
API supports a stack of popups.

This commit converts all client use of the old API to the newer API;
some places still call the older APIs, however, in order to behave
correctly if an older version of the game is being loaded.