Add HTML scroller
This adds a function that wraps the contents of a div into a scrollable
container.

 - default scrollbars are replaced (they look seriously out of place)
 - semitransparent overflow 'shades' are added at the top and bottom of
   the scroll container.

In addition, we place the scroll container on its own render layer.
Effectively this means that the menu contents are rendered to a texture,
rather than rendered directly; this means that scrolling is much faster,
as painting now consists of translating and rendering that texture,
as opposed to doing a layout pass over the menu items. With large menus
(the largest I know of is ?/m .*) this takes menu two-finger scrolling
from very janky to perfectly smooth.